fixed t01, t03

+++ src/main/cpp/t01_frac.cpp
@@ -20,7 +20,7 @@
 using namespace std;
 
 int t01_frac() {
-    float inp, whole;
+    double inp, whole;
     cin >> inp;
     whole = floor(inp);
     inp = inp - whole;
